Adding the Pokemon Detail

+++ app/src/main/java/martinez/andres/modulo6practica2/data/model/Pokemon.kt
@@ -3,7 +3,7 @@ package martinez.andres.modulo6practica2.data.model
 import com.google.gson.annotations.SerializedName
 
 data class Pokemon(
-    @SerializedName("id") val id: Long,
+    @SerializedName("id") val id: Int,
     @SerializedName("name") val name: String,
     @SerializedName("urlImage") val urlImage: String,
     @SerializedName("description") val description: String?,
